Add handle exceptions decorator
