+++ binder/view_logger.py
@@ -1,52 +1,15 @@
-import time
-import django
-from django.http.request import RawPostDataException
+import warnings
 
-from binder.views import ellipsize
+from .decorators import view_logger
 
 
-def view_logger(logger, log_request_body=True):
-	def decorator(view):
-		def res(request, *args, **kwargs):
-			time_start = time.time()
-			logger.info('request dispatch; verb={}, user={}/{}, path={}'.format(
-				request.method,
-				request.user.id,
-				request.user,
-				request.path,
-			))
-			logger.info('remote_addr={}, X-Real-IP={}, X-Forwarded-For={}'.format(
-				request.META.get('REMOTE_ADDR', None),
-				request.META.get('HTTP_X_REAL_IP', None),
-				request.META.get('HTTP_X_FORWARDED_FOR', None),
-			))
-			logger.info('request parameters: {}'.format(dict(request.GET)))
-			logger.debug('cookies: {}'.format(request.COOKIES))
+warnings.warn(
+	(
+		'importing view_logger from binder.view_logger is deprecated, '
+		'view_logger should be imported from binder.decorators'
+	),
+	DeprecationWarning,
+	stacklevel=2,
+)
 
-			if not log_request_body:
-				body = ' censored.'
-			else:
-				# FIXME: ugly workaround, remove when Django bug fixed
-				# Try/except because https://code.djangoproject.com/ticket/27005
-				try:
-					if request.META.get('CONTENT_TYPE', '').lower() == 'application/json':
-						body = ': ' + ellipsize(request.body, length=65536)
-					else:
-						body = ': ' + ellipsize(request.body, length=64)
-				except RawPostDataException:
-					body = ' unavailable.'
-
-			logger.debug('body (content-type={}){}'.format(request.META.get('CONTENT_TYPE'), body))
-
-			response = view(request, *args, **kwargs)
-
-			logger.info('request response; status={} time={}ms bytes={} queries={}'.format(
-				response.status_code,
-				int((time.time() - time_start) * 1000),
-				'?' if response.streaming else len(response.content),
-				len(django.db.connection.queries),
-			))
-
-			return response
-		return res
-	return decorator
+__all__ = ['view_logger']
